Fix: Apply automated formatting and fix build warnings

+++ development/include/GXDLMSSha1.h
@@ -38,14 +38,13 @@
 #include "GXBytebuffer.h"
 
 //This class is used to handle SHA-1.
-class CGXDLMSSha1
-{
+class CGXDLMSSha1 {
 private:
-    static void Transform(unsigned long *block, unsigned int *digest, unsigned int *transforms);
-    static void Update(CGXByteBuffer& data, unsigned int *digest, unsigned int *transforms);
-    static int Final(CGXByteBuffer& data, unsigned int *digest, unsigned int *transforms, CGXByteBuffer& reply);
+    static void Transform(uint32_t *block, unsigned int *digest, unsigned int *transforms);
+    static void Update(CGXByteBuffer &data, unsigned int *digest, unsigned int *transforms);
+    static int Final(CGXByteBuffer &data, unsigned int *digest, unsigned int *transforms, CGXByteBuffer &reply);
 
 public:
-    static int Encrypt(CGXByteBuffer& data, CGXByteBuffer& crypted);
+    static int Encrypt(CGXByteBuffer &data, CGXByteBuffer &crypted);
 };
-#endif //GXDLMSSHA1_H
+#endif  //GXDLMSSHA1_H
